requirements.txt
streamlit
torch
torchvision
Pillow
numpy
openai
tensorflow
pandas
nltk
